Lots of things. Better GameEntity. Player, Invader, etc.
Currently doesn't actually draw the player even though it should be
getting a valid texture. Hrm.

+++ src/main/texture/Texture.java
@@ -21,8 +21,8 @@ public abstract class Texture {
     public abstract float getTexTopLeftY();
     public abstract float getTexBottomLeftY();
 
-    public abstract double getWidth();
-    public abstract double getHeight();
+    public abstract float getWidth();
+    public abstract float getHeight();
 
     public abstract int getName();
     public abstract int getTarget();
